Use a musical measure instead of time
We want to send an event every beat at a given tempo - introducing
this struct lets us use more domain-specific language.

I'm having to do a lot of type conversions in `to_ms` - this might not
be very idiomatic? I want to return a `u64` because that's what
`Duration::from_millis` takes.