Replace hard-coded rhythm with a pattern
This commit allows me to shorten the tick length in the main loop and
still maintain the kick on every beat of the bar.

The pattern struct currently assumes that the pattern is a single bar
in 4/4 time.

The elements of the pattern vector are start times for the notes in
the pattern. At the moment each note can only be the kick as defined
in the Event struct but we can replace that with a richer type later.

I've chosen to have the tick increase monotonically forever and then
to reckon this into the repetition within the `notes_between` method
using the `reduce_to_one_bar` method on the Measure. I'm not sure if I
like this yet, but we'll see how it shakes out.