Rename pattern -> track
I was reading an old MPC manual[1] and they use "track" to represent a
collection of notes, while a Sequence is a collection of tracks. I
think what we have at the moment is an implicit a 1-bar sequence that
loops, and contains a single track with that track having 4 events
triggering the same sound.

[1] https://www.mpc-tutor.com/files/mpc-manual/MPC60_manual_v310.pdf