Subtract processing time from sleep time
It takes a varying amount of time to perform the work in each `loop`,
so we shouldn't really be sleeping for the same amount of time at the
end of each loop. Instead we subtract the processing time from the
ideal length of each tick to compensate.

I think we could do something more sophisticated here and use the
`loop_counter` to work out how much time should have passed to account
for any drift, but perhaps we can revisit that later.

Ideally, I'd like some tooling to measure how stable this timer is to
know whether we need to improve things further.