Add step to track by hitting atom pad
When I hit pad 1 one in bank 1 on the presonus atom `m2o` sends a
value in the OSC message argument of `24` - which is parsed as
`Int(36)` by OSC (24 is 36 in hexadecimal). Similarly pad 16 is `33`
in hex, or 51 in decimal.

This change translates the incoming messages in OSC's format to add
the appropriate Measure to the track.