Ignore note off messages when toggling steps
Before this commit we were calling `toggle_step` twice each time a
controller pad was hit. This was because hitting a pad sends both a
`note_on` and a `note_off` message, and our message handling was being
too liberal in accepting both of these as a `note_on`.

I've decided to explicitly parse the NoteOff message even though we
don't need it at the moment, as it makes it a bit clearer what's going
on, I think.