First pass at updating atom lights
I'm trying to avoid having to store messages about which lights need
to be turned on or off. Instead I've tried to turn all lights off and
then just the lights for the current step on in each iteration of the
loop. Unfortunately that seems like too many messages to be sending,
as the Atom lights flicker quite a lot.

Instead I'm going to try and work out the difference between each
successive iteration to minimise the amount of messages that need to
be sent.