Store a vector of Measure rather than Events on Track
I think the active steps for a given track could be modelled better,
perhaps by using a set rather than a vector (which would allow me to
calculate the difference in active steps between two tracks more
easily).

As a first step, I've changed `steps` to be a `Vec<Measure>` and
isolated the conversion to events for the current time window to the
`events_between` function.