Only turn on/off pads that have changed in each update
In cde1e73 I had a very simple method for updating the Atom pad lights
which did not perform well. Now that we can more easily get the
difference between one tick and the next (using the `difference`
function on the HashSet that stores the active steps) we can work out
which lights to turn off, and which to turn on.

I'd like to write a test for this, but I'm not sure how to. Perhaps I
need to use dependency injection to be able to mock out the
interaction over OSC. I'll "come back to this" later.