Allow multiple samples to be played by the sampler
I've changed the received messages format from

    /sampler/1

to

   /sampler <i>

Where the integer argument plays the equivalently named file in the
`samples/` directory (i.e. `/sampler 1` will play
`samples/1.wav`. This allows me to map midi note numbers to samples
fairly easily in the first instance.

I've initialized the dictionary which stores the mapping of file name
to buffer to take 16 keys for now. That's sufficient for one sample on
each pad of the atom controller.