Store the active note number in the context
We want to allow the steps for an individual note to be sequenced. To
do that we'll store an "active note" on the Context, so that the note
to be sequenced can be changed. Currently this is hard-coded for note
number 1, but we will shortly allow that to be changed via MIDI/OSC message.