Fix bug in atom::update()
Before this commit I was comparing Steps between the current context
and the next to determine which lights to turn on or off. If a active
note number had the same pad lit as the next note we were changing
too, we were incorrectly turning that light off, because the
`difference` was based on the Step (which includes the
note_number). Instead we reduce the active steps down to just the
lights that are illuminated for the current active step and this fixes
the problem.

This function doesn't have any tests, which helped the bug to creep
in :(.