Improve tests around atom::update()
This extracts the main logic for this function into a separate
function and tests it. It's a bit harder to test the full update()
function because it has side-effects (sending OSC messages) rather
than a return value.

This does make me think that this module should be returning OSC
messages to be sent by the outer shell.