Replace active_{steps,sixteenths}_with_note_number
The Atom UI implicitly assumes that the steps toggled are sixteenth
notes in a one-bar pattern (there's sixteen pads on the atom and we
don't have a way to switch between bars).

I'm trying to avoid exposing Step publically so making this change
removes one place it is currently exposed.