Improve testing around events_for_tick and fix off-by-one error
Musically we want to count "1 and 2 and 3 and 4", so we'd expect notes
on the "1st sixteenth" to fire at the start of the bar. The start of
the bar is offset 0 into the track, and the tick counter counts 0, 1,
2, ...

These changes make the testing around this behaviour more robust and
fix an off-by-one error that was causing a pause at the end of each repeat.