Implement swing
The swing amount is expressed in ticks and applies to all of the
events generated by the track. If an event falls on an even sixteenth
and the swing amount is more than 0 then the event is delayed by that
many ticks.

This is my interpretation of how the swing was implemented by Roger
Linn from an interview[1] with him in Attack Magazine.

[1] https://www.attackmagazine.com/features/interview/roger-linn-swing-groove-magic-mpc-timing/