Set swing amount using KnobIncrement / KnobDecrement
If we directly increment the swing amount by 1 tick each time we
receive a KnobIncrement message, there's not enough range on the knob
to fine tune the amount. Instead we store the swing amount as a
percentage (0 to the maximum amount of swing 100) and convert that to
ticks when we apply the swing.

I've chose 6 ticks to be the maximum amount of swing, because at 96
pulses per 4-beat-measure, 6 ticks means that the even 16th note is
swung so that it lies on top of the next odd sixteenth, that-is the
maximum amount of swing we can apply between two sixteenths.