Output MIDI clock
This sends the message `/*/clock` to `o2m` which broadcasts the MIDI
clock message to all devices.

As far as I can tell, MIDI clock should be sent 24 times per quarter
note, which is the same as the main event loop runs at (see
`main::tick_duration()`).

I've tested this with my Volca Keys and it seems to do the right
thing.

By default on OS X broadcasting to all channels also broadcasts to the
IAC (inter-application communication) MIDI driver which will therefore
send a clock to any running MIDI application. For now, that seems ok -
it is possible to specify output devices as `o2m` command line args,
so that would be a work-around for now, and avoids having to add
additional config to this app.

Closes #6