Allow controller name to be passed via CLI args
o2m uses the device name in its message format. This is because a new
MIDI port is created based on the name of the device when the device
is connected by USB. For the Presonus Atom controller I am using, this
name is different by default in OS X (atom) compared to using ALSA on
the Raspberry Pi (atom_midi_1). This means that before this commit we
couldn't send messages to the Atom from the Pi.

I imagine there's probably some kind of way of telling the OS to use a
different name for the MIDI port it creates for a given device, but I
think it's probably easier to have a default, and allow it to be
configured.

I've chosen to use the Clap crate and command line arguments for
this. I've taken `atom` to be the default, which means the name will
need passing on the Pi.

Closes #8