Pass a reference to the UdpSocket to process_incoming_message
Before this commit, processing incoming messages was broken. In
6689a12 I'd moved this logic into a new module, but put the creation
of the UdpSocket inside the function that was called every time the
loop ran. This meant that the socket couldn't be created on every
iteration of the loop and wasn't being bound too. Passing a reference
to the socket fixes that issue.