Add start and stop commands at the top level of the directory
This allows all of the services to be started with a single
command. I'm parsing out the name of the atom controller from o2m to
allow this code to run both on OS X and on the Pi.

The sleep is a bit of a hack to increase the chance that m2o and o2m
have started before we start the sequencer (since it needs to output
to the controller when it starts up).