Prevent panic if pulses is reduced too far.
Pulses always needs to be greater than or equal to the number of
onsets. It also needs to be greater than or equal to rotate due to how
it's implemented using `rotate_right`. It also can't be 0 (or less)
because we divide by it to calculate `offset_into_pattern`. So when
it's decremented we also decrement `onsets` and `rotate` if they'd be
larger, and we clamp it to >= 1.