Introduce Step mode
This mode is intended to be a page which allows us to chose which note
in the sequence to edit. For now, I've made it so that it doesn't
return a `current_pad` (so no indication of progress through the
pattern), because the pads won't be associated with time in this
view.

I've also returned an empty set for `active_pads` for now, we can
think about something useful to do here later.