Add one shot sounds to the step sequence
When we hit a pad in Step mode, we add that note to a vector of "one
shots" for the next tick. This means the note is immediately
played (so we can finger drum in Step mode).

When we add a new one shot, we clear out any old one shots as we don't
have any use for them at the moment (and we don't want the vector to
grow without bounds).

Clear old one shots when we add new ones

Extract OneShotSequencer from StepSequencer