Flash the <select> button when in StepSelect mode
This is a slightly lazy solution, rather than use the same difference
update mechanism as for the other buttons, I've decided to just send
either a "on" or "off" message every 6 ticks (or an "off" message
every 12 ticks when not in StepSelect mode).

We change the light on the <select> button by sending a MIDI Control
Change message on channel 1 to controller channel 103.

Closes #24