Introduce Sequence
This type is intended to cover the functionality of both the existing
StepSequencer and EuclideanSequencer. I want a more abstract concept
of a sequence which consists of steps, and note-on triggers (which can
be shifted within the step by an offset). This commit allows for
triggers to be added to the sequence and returned for a given tick.