Implement euclidean_fill() for Sequence
This function returns a new Sequence of the same length as `self` with
steps filled using the euclidean algorithm (as first implemented in
137c0a6fb).

Although each step can have multiple triggers for different note
numbers, I've assumed for now that if we call euclidean_fill() we want
to create a new sequence with a euclidean pattern for a single note
number. This will be the case when a Sequence represents a drum part,
for example.