Immediately play default note when in performance mode
This commit allows the default note for a given sequence to be
triggered immediately (as in, at the next tick) when the corresponding
pad is hit in performance mode. This makes it much easier to audition
sounds when composing and to trigger one-shot samples during a
performance.

The `performance_events` vector only stores events to be played on the
next tick of the main loop, so we clear it after we have triggered the
events it contains.